// frontend/src/features/user-info/types/user.ts
/**
 * 유저 기본 정보 타입 정의
 */

export interface UserInfo {
	character_name: string;
	world_name: string;
	character_level: number;
	character_gender: string;
	character_class: string;
	character_class_level: string;
	character_exp: number;
	character_exp_rate: string;
	character_guild_name: string;
	character_image: string;
	character_date_create: string;
}